refactor(db): explicitly acquire and release db connections
This is necessary as virtual-connection? only releases connections
back to the pool when the handling thread gets closed, but web server
handler threads only get closed when the associated browser connection
is closed (assuming keep-alive is enabled, which we want), meaning
that connections can be kept out of the pool without being used for
arbitrarily long periods of time and that's no good!